make it compilable with win64

+++ src/KSPACE/msm.cpp
@@ -279,7 +279,7 @@ double MSM::estimate_cutoff(double h, double prd)
   a = C_p*pow(h,(p-1))/accuracy;
 
   // include dependency of error on other terms
-  a *= q2/(prd*sqrt(atom->natoms));
+  a *= q2/(prd*sqrt(double(atom->natoms)));
 
   a = pow(a,1.0/double(p));
 
@@ -330,7 +330,7 @@ double MSM::estimate_1d_error(double h, double prd)
   double error_1d = C_p*pow(h,(p-1))/pow(a,(p+1));
 
   // include dependency of error on other terms
-  error_1d *= q2*a/(prd*sqrt(atom->natoms));
+  error_1d *= q2*a/(prd*sqrt(double(atom->natoms)));
 
   return error_1d;
 }
